Refactor grammar
